+++ membership.php
@@ -5,7 +5,7 @@
  * @author      Lee Garner <lee@leegarner.com>
  * @copyright   Copyright (c) 2012-2020 Lee Garner <lee@leegarner.com>
  * @package     membership
- * @version     v0.2.0
+ * @version     v0.2.2
  * @license     http://opensource.org/licenses/gpl-2.0.php
  *              GNU Public License v2 or later
  * @filesource
@@ -38,7 +38,7 @@ $_TABLES['membership_users']    = $prefix . 'users';
  */
 global $_CONF_MEMBERSHIP;
 $_CONF_MEMBERSHIP['pi_name']           = 'membership';
-$_CONF_MEMBERSHIP['pi_version']        = '0.2.1';
+$_CONF_MEMBERSHIP['pi_version']        = '0.2.2';
 $_CONF_MEMBERSHIP['gl_version']        = '1.7.8';
 $_CONF_MEMBERSHIP['pi_url']            = 'http://www.leegarner.com';
 $_CONF_MEMBERSHIP['pi_display_name']   = 'Membership';