+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Release TBD
 - Use Uikit datepicker
 - Change Paypal plugin to Shop
 - Add a message to plan lists indicating the latest plan if expired.
+- Positions can be optionally shown on profile lists.
 
 ## 0.2.0
 - Implement PHP class autoloader